refactor(tests): split test-fixtures into folders with chained .extend() builder
Each fixture group now lives in its own module (addresses, relayer, signer,
provider, storage, services, sdk, token for the SDK; token, query-client,
wrapper for react-sdk) and exports a typed fixtures object. The package index
composes them via vitest's `test.extend(A).extend(B)...` builder chain so
dependencies between groups are explicit and reading order matches resolution
order. Node folder resolution keeps `from "../../test-fixtures"` working
unchanged across all 125+ consumer test files.
