+++ package.json
@@ -78,6 +78,7 @@
     "@types/react": "19.2.14",
     "@types/react-dom": "19.2.3",
     "@vitest/coverage-v8": "^4.1.5",
+    "@vitest/runner": "^4.1.6",
     "@vitest/ui": "^4.1.5",
     "@zama-fhe/relayer-sdk": "0.4.2",
     "conventional-changelog-conventionalcommits": "^9.3.1",
