+++ packages/react-sdk/src/test-fixtures/constants.ts
@@ -1,0 +1,2 @@
+export const TEST_ADDR_A = "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAa" as const;
+export const TEST_ADDR_B = "0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB" as const;
