+++ packages/sdk/src/test-fixtures/sdk.ts
@@ -1,0 +1,56 @@
+// oxlint-disable eslint-plugin-react-hooks/rules-of-hooks
+import type { ZamaConfig } from "../config/types";
+import { ZamaSDKEvents } from "../events/sdk-events";
+import type { RelayerSDK } from "../relayer/relayer-sdk";
+import type { GenericProvider, GenericSigner, GenericStorage } from "../types";
+import { ZamaSDK } from "../zama-sdk";
+import { createMockChain } from "./chain";
+import type { ProviderFixtures } from "./provider";
+import type { RelayerFixtures } from "./relayer";
+import type { SignerFixtures } from "./signer";
+import type { StorageFixtures } from "./storage";
+import type { FixturesOf } from "./types";
+
+export type CreateSDKFn = (overrides?: Partial<ZamaConfig>) => ZamaSDK;
+
+export interface SdkFixtures {
+  sdk: ZamaSDK;
+  createSDK: CreateSDKFn;
+  events: typeof ZamaSDKEvents;
+}
+
+function buildSDK(
+  relayer: RelayerSDK,
+  provider: GenericProvider,
+  signer: GenericSigner,
+  storage: GenericStorage,
+  overrides?: Partial<ZamaConfig>,
+): ZamaSDK {
+  return new ZamaSDK({
+    chains: [createMockChain({ id: 31337 })],
+    relayer: relayer as unknown as ZamaConfig["relayer"],
+    provider,
+    signer,
+    storage,
+    permitStorage: storage,
+    keypairTTL: 2592000,
+    permitTTL: 1,
+    registryTTL: 86400,
+    onEvent: undefined,
+    ...overrides,
+  } as unknown as ZamaConfig);
+}
+
+type SdkDeps = RelayerFixtures & ProviderFixtures & SignerFixtures & StorageFixtures;
+
+export const sdkFixtures: FixturesOf<SdkFixtures, SdkDeps> = {
+  sdk: async ({ relayer, provider, signer, storage }, use) => {
+    await use(buildSDK(relayer, provider, signer, storage));
+  },
+  createSDK: async ({ provider, signer, relayer, storage }, use) => {
+    const factory: CreateSDKFn = (overrides) =>
+      buildSDK(relayer, provider, signer, storage, overrides);
+    await use(factory);
+  },
+  events: ZamaSDKEvents,
+};
