+++ packages/sdk/tsconfig.json
@@ -8,7 +8,6 @@
     "src/**/*.test.ts",
     "src/**/*.test.tsx",
     "src/**/__tests__/**",
-    "src/test-fixtures.ts",
     "dist",
     "node_modules"
   ]
